+++ lib/elixir/test/elixir/list_test.exs
@@ -126,30 +126,30 @@ defmodule ListTest do
   end
 
   test :keyfind do
-    assert List.keyfind([a: 1, b: 2], :a, 1) == { :a, 1 }
-    assert List.keyfind([a: 1, b: 2], 2, 2) == { :b, 2 }
-    assert List.keyfind([a: 1, b: 2], :c, 1) == nil
+    assert List.keyfind([a: 1, b: 2], :a, 0) == { :a, 1 }
+    assert List.keyfind([a: 1, b: 2], 2, 1) == { :b, 2 }
+    assert List.keyfind([a: 1, b: 2], :c, 0) == nil
   end
 
   test :keyreplace do
-    assert List.keyreplace([a: 1, b: 2], :a, 1, { :a, 3 }) == [a: 3, b: 2]
-    assert List.keyreplace([a: 1], :b, 1, { :b, 2 }) == [a: 1]
+    assert List.keyreplace([a: 1, b: 2], :a, 0, { :a, 3 }) == [a: 3, b: 2]
+    assert List.keyreplace([a: 1], :b, 0, { :b, 2 }) == [a: 1]
   end
 
   test :keystore do
-    assert List.keystore([a: 1, b: 2], :a, 1, { :a, 3 }) == [a: 3, b: 2]
-    assert List.keystore([a: 1], :b, 1, { :b, 2 }) == [a: 1, b: 2]
+    assert List.keystore([a: 1, b: 2], :a, 0, { :a, 3 }) == [a: 3, b: 2]
+    assert List.keystore([a: 1], :b, 0, { :b, 2 }) == [a: 1, b: 2]
   end
 
   test :keymember? do
-    assert List.keymember?([a: 1, b: 2], :a, 1) == true
-    assert List.keymember?([a: 1, b: 2], 2, 2) == true
-    assert List.keymember?([a: 1, b: 2], :c, 1) == false
+    assert List.keymember?([a: 1, b: 2], :a, 0) == true
+    assert List.keymember?([a: 1, b: 2], 2, 1) == true
+    assert List.keymember?([a: 1, b: 2], :c, 0) == false
   end
 
   test :keydelete do
-    assert List.keydelete([a: 1, b: 2], :a, 1) == [{ :b, 2 }]
-    assert List.keydelete([a: 1, b: 2], 2, 2) == [{ :a, 1 }]
-    assert List.keydelete([a: 1, b: 2], :c, 1) == [{ :a, 1 }, { :b, 2 }]
+    assert List.keydelete([a: 1, b: 2], :a, 0) == [{ :b, 2 }]
+    assert List.keydelete([a: 1, b: 2], 2, 1) == [{ :a, 1 }]
+    assert List.keydelete([a: 1, b: 2], :c, 0) == [{ :a, 1 }, { :b, 2 }]
   end
 end