+++ lib/elixir/test/elixir/tuple_test.exs
@@ -4,10 +4,10 @@ defmodule TupleTest do
   use ExUnit.Case, async: true
 
   test :elem do
-    assert elem({ :a, :b, :c }, 2) == :b
+    assert elem({ :a, :b, :c }, 1) == :b
   end
 
   test :setelem do
-    assert setelem({ :a, :b, :c }, 2, :d) == { :a, :d, :c }
+    assert setelem({ :a, :b, :c }, 1, :d) == { :a, :d, :c }
   end
 end