+++ lib/mix/lib/mix/task.ex
@@ -60,7 +60,7 @@ defmodule Mix.Task do
   Returns the shortdoc or `nil`.
   """
   def shortdoc(module) when is_atom(module) do
-    case List.keyfind module.__info__(:attributes), :shortdoc, 1 do
+    case List.keyfind module.__info__(:attributes), :shortdoc, 0 do
       { :shortdoc, [shortdoc] } -> shortdoc
       _ -> nil
     end
@@ -70,7 +70,7 @@ defmodule Mix.Task do
   Checks if the task is hidden or not. Returns a boolean.
   """
   def hidden?(module) when is_atom(module) do
-    case List.keyfind module.__info__(:attributes), :hidden, 1 do
+    case List.keyfind module.__info__(:attributes), :hidden, 0 do
       { :hidden, [bool] } -> bool
       _ -> false
     end